+++ src/Microsoft.SqlTools.ServiceLayer/Connection/ConnectionInfo.cs
@@ -12,6 +12,7 @@ using System.Collections.Generic;
 using System.Data.Common;
 using System.Threading.Tasks;
 using Microsoft.SqlServer.Management.Common;
+using Microsoft.SqlTools.LanguageService.LanguageServices;
 using Microsoft.SqlTools.ServiceLayer.Connection.Contracts;
 using Microsoft.SqlTools.SqlCore.Connection;
 using Microsoft.SqlTools.Utility;
